Implement more CommonEconomy

+++ totaleconomy-common/src/main/java/com/ericgrandt/totaleconomy/common/data/BalanceData.java
@@ -65,7 +65,7 @@ public class BalanceData {
         }
     }
 
-    public void transfer(UUID fromAccountId, UUID toAccountId, int currencyId, double amount) throws SQLException {
+    public void transfer(UUID fromAccountId, UUID toAccountId, int currencyId, BigDecimal amount) throws SQLException {
         String withdrawQuery = "UPDATE te_balance SET balance = balance - ? WHERE account_id = ? AND currency_id = ?";
         String depositQuery = "UPDATE te_balance SET balance = balance + ? WHERE account_id = ? AND currency_id = ?";
 
@@ -78,12 +78,12 @@ public class BalanceData {
                 PreparedStatement withdrawStmt = conn.prepareStatement(withdrawQuery);
                 PreparedStatement updateStmt = conn.prepareStatement(depositQuery)
             ) {
-                withdrawStmt.setBigDecimal(1, BigDecimal.valueOf(amount));
+                withdrawStmt.setBigDecimal(1, amount);
                 withdrawStmt.setString(2, fromAccountId.toString());
                 withdrawStmt.setInt(3, currencyId);
                 withdrawStmt.executeUpdate();
 
-                updateStmt.setBigDecimal(1, BigDecimal.valueOf(amount));
+                updateStmt.setBigDecimal(1, amount);
                 updateStmt.setString(2, toAccountId.toString());
                 updateStmt.setInt(3, currencyId);
                 updateStmt.executeUpdate();
